Write full tensors out at once in HF consolidation script
Pull Request resolved: https://github.com/pytorch/pytorch/pull/159394

Not all storage systems support writing at random offsets. This PR changes the writes of the consolidation script to write each tensor to a buffer, and then write out the buffer, sequentially going through every tensor in the output file. This will also help in the case where the sharded files weren't just sharded in the row-wise dimension. The reason is because small writes are expensive and we were writing each write for every chunk that was the largest number of contiguous bytes in the final tensor, but this could be a small amount of bytes for col-wise sharding. Now the full tensor is needed for the write, making the number of small writes smaller.
ghstack-source-id: 300234774
@exported-using-ghexport

Differential Revision: [D78684452](https://our.internmc.facebook.com/intern/diff/D78684452/)