[OpenReg] Add Develop Notes for Integrating New Backend into PyTorch
As the title stated.

To facilitate the integration of the new backend, we plan to publish a new
development note that details all the key components, hoping to speed up
the development of other accelerators.

This PR is the beginning of this note, which we will gradually improve and
keep in sync with OpenReg's code.

ghstack-source-id: c7863226283229b88b0e7cd1c953133479807fef
Pull-Request-resolved: https://github.com/pytorch/pytorch/pull/158644